+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * [#3670](https://github.com/bbatsov/rubocop/pull/3670): Add `CountBlocks` boolean option to `Metrics/BlockNesting`. It allows blocks to be counted towards the nesting limit. ([@georgyangelov][])
 * [#2992](https://github.com/bbatsov/rubocop/issues/2992): Add a configuration to `Style/ConditionalAssignment` to toggle offenses for ternary expressions. ([@rrosenblum][])
 * [#3824](https://github.com/bbatsov/rubocop/pull/3824): Add new `Perfomance/RegexpMatch` cop. ([@pocke][])
+* [#3825](https://github.com/bbatsov/rubocop/pull/3825): Add new `Rails/SkipsModelValidations` cop. ([@rahulcs][])
 
 ### Changes
 
@@ -2567,3 +2568,4 @@
 [@enriikke]: https://github.com/enriikke
 [@iguchi1124]: https://github.com/iguchi1124
 [@vergenzt]: https://github.com/vergenzt
+[@rahulcs]: https://github.com/rahulcs